Moved CDR interfaces to its own folder
