+++ tgui/packages/tgui/interfaces/Cdr/AtmosCdr.tsx
@@ -1,6 +1,6 @@
 import { BooleanLike } from 'common/react';
 
-import { useBackend } from '../backend';
+import { useBackend } from '../../backend';
 import {
   Box,
   Button,
@@ -13,8 +13,8 @@ import {
   NumberInput,
   ProgressBar,
   Section,
-} from '../components';
-import { Window } from '../layouts';
+} from '../../components';
+import { Window } from '../../layouts';
 
 type CoreComposition = Record<string, number>;
