+++ tgui/packages/tgui/interfaces/Cdr/Graph.tsx
